[libc++] Avoid more <format> code in <vector>
This makes it ~45% faster to parse `<vector>` on my system.
