+++ libcxx/include/__format/formatter_bool.h
@@ -13,7 +13,7 @@
 #include <__config>
 #include <__format/concepts.h>
 #include <__format/formatter.h>
-#include <__format/parser_std_format_spec.h>
+#include <__format/parser_std_format_spec_data.h>
 
 #if !defined(_LIBCPP_HAS_NO_PRAGMA_SYSTEM_HEADER)
 #  pragma GCC system_header
@@ -23,17 +23,19 @@ _LIBCPP_BEGIN_NAMESPACE_STD
 
 #if _LIBCPP_STD_VER >= 20
 
+template <__fmt_char_type _CharT, class _ParseContext>
+_LIBCPP_HIDE_FROM_ABI constexpr
+    typename _ParseContext::iterator __formatter_bool_parse(__format_spec::__parser_data<_CharT>&, _ParseContext&);
+
 template <__fmt_char_type _CharT, class _FormatContext>
 _LIBCPP_HIDE_FROM_ABI typename _FormatContext::iterator
-__formatter_bool_format(bool __value, __format_spec::__parser<_CharT>, _FormatContext&);
+__formatter_bool_format(bool __value, __format_spec::__parser_data<_CharT>, _FormatContext&);
 
 template <__fmt_char_type _CharT>
 struct formatter<bool, _CharT> {
   template <class _ParseContext>
   _LIBCPP_HIDE_FROM_ABI constexpr typename _ParseContext::iterator parse(_ParseContext& __ctx) {
-    typename _ParseContext::iterator __result = __parser_.__parse(__ctx, __format_spec::__fields_integral);
-    __format_spec::__process_parsed_bool(__parser_, "a bool");
-    return __result;
+    return std::__formatter_bool_parse(__parser_, __ctx);
   }
 
   template <class _FormatContext>
@@ -41,7 +43,7 @@ struct formatter<bool, _CharT> {
     return std::__formatter_bool_format(__value, __parser_, __ctx);
   }
 
-  __format_spec::__parser<_CharT> __parser_;
+  __format_spec::__parser_data<_CharT> __parser_;
 };
 
 #  if _LIBCPP_STD_VER >= 23
