+++ libcxx/include/__format/range_formatter.h
@@ -207,7 +207,8 @@ struct range_formatter {
     return ranges::copy(__closing_bracket_, __ctx.out()).out;
   }
 
-  __format_spec::__parser<_CharT> __parser_{.__alignment_ = __format_spec::__alignment::__left};
+  __format_spec::__parser<_CharT> __parser_ = {
+      __format_spec::__parser_data<_CharT>{.__alignment_ = __format_spec::__alignment::__left}};
 
 private:
   template <contiguous_iterator _Iterator>
