+++ libcxx/test/libcxx/transitive_includes/cxx26.csv
@@ -741,7 +741,6 @@ queue initializer_list
 queue iosfwd
 queue limits
 queue stdexcept
-queue string
 queue string_view
 queue tuple
 queue vector
@@ -1114,7 +1113,6 @@ vector initializer_list
 vector iosfwd
 vector limits
 vector stdexcept
-vector string
 vector string_view
 vector tuple
 vector version
